groundwork is laid for graphql api. TODO: write resolvers
