+++ package.json
@@ -39,8 +39,6 @@
     "react-dom": "^19.0.0"
   },
   "devDependencies": {
-    "@eddeee888/gcg-typescript-resolver-files": "^0.12.0",
-    "@graphql-codegen/cli": "5.0.5",
     "@stylistic/eslint-plugin": "^4.2.0",
     "@types/node": "^22.13.9",
     "@types/react": "^19.0.10",
@@ -54,10 +52,7 @@
     "pino-pretty": "^13.0.0",
     "postcss": "^8.5.3",
     "tailwindcss": "^3.4.17",
-    "typescript": "^5.8.2",
-    "@graphql-codegen/typescript": "4.1.5",
-    "@graphql-codegen/typescript-document-nodes": "4.0.15",
-    "@graphql-codegen/introspection": "4.0.3"
+    "typescript": "^5.8.2"
   },
   "overrides": {
     "uri-js": "pnpm:uri-js-replace"
