--- src/lib/graphql/schema.graphql
@@ -1,28 +1,0 @@
-type Query {
-	tag(name: String!): Tag!
-	resource(name: String!): Resource!
-}
-
-type Tag {
-	name: String!
-	description: String
-	resources: [Resource!]!
-}
-
-type Resource {
-	name: String!
-	description: String
-	URL: String!
-	tags: [Tag!]!
-}
-
-type TagMapper = {
-	tagName: String
-	tagDescription: String
-}
-
-type ResourceMapper = {
-	resName: String
-	resDescription: String
-	resURL: String
-}
