+++ src/lib/graphql/typeDefs.ts
@@ -1,0 +1,20 @@
+const typeDefs: any = /*GraphQL */ `
+	type Query {
+		tag(name: String!): Tag!
+	}
+
+	type Tag {
+		name: String!
+		description: String
+		resources: [Resource!]!
+	}
+
+	type Resource {
+		name: String!
+		description: String
+		URL: String!
+		tags: [Tag!]!
+	}
+`;
+
+export default typeDefs;
